Updates for RHEL/CentOS 7.2 updates

+++ recipes/base_install.rb
@@ -48,7 +48,7 @@ python_package 'awscli'
 
 # TODO: update nfs receipes to stop, disable nfs services
 include_recipe "nfs"
-if node['platform_family'] == 'rhel' && node['platform_version'].to_f >= 7.0 && node['platform'] != 'amazon'
+if node['platform_family'] == 'rhel' && node['platform_version'].to_i >= 7 && node['platform'] != 'amazon'
   service "rpcbind" do
     action [:start, :enable]
     supports status: true
